extract styled primitive out
