+++ packages/forms/src/components/SwitchInput/styles.ts
