+++ packages/ui/src/index.tsx
@@ -63,6 +63,7 @@ import Sidebar, { SidebarProps } from "./components/Sidebar";
 import Spoiler, { SpoilerProps } from "./components/Spoiler/Spoiler";
 import Stepper, { StepperProps } from "./components/Stepper/Stepper";
 import { StepType } from "./components/Stepper/types";
+import Switch, { SwitchProps } from "./components/Switch/Switch";
 import TableOfContents, {
   type TocAnchorLinkProps,
   TocHeadingProps,
@@ -111,6 +112,7 @@ export type {
   SpoilerProps,
   StepperProps,
   StepType,
+  SwitchProps,
   TocAnchorLinkProps,
   TocHeadingProps,
   TocSectionProps,
@@ -160,6 +162,7 @@ export {
   SideMenuContentWrapper,
   Spoiler,
   Stepper,
+  Switch,
   TableOfContents,
   Tabs,
   ToggleGroup,
